added+integrated candidate folder (DOES NOT BUILD, SEE ISSUE IN "apps/reviewers/models.py")
