+++ mayan/settings/base.py
@@ -117,6 +117,7 @@ INSTALLED_APPS = (
     'mayan.apps.document_signatures',
     'mayan.apps.document_states',
     'mayan.apps.documents',
+    'mayan.apps.candidates',
     'mayan.apps.reviewers',
     'mayan.apps.duplicates',
     'mayan.apps.file_metadata',
